refactor: Migrate data to pyproject.toml (#22)
This PR removes setup.py and setup.cfg in favor using a fully PEP
517-compliant build. The linter configs have also moved to
pyproject.toml where possible. flake8 does not (and has no plans to)
support pyproject.toml, and since it's the only relevant tool doing that
I migrated its configuration to a flake8-specific file. I also renamed
CONTRIBUTORS.md to CONTRIBUTING.md to better describe what it is
(instructions for contributors, not a list of contributors).